Handle Http 300-range codes in response checker
Now anything <300 is regarded as success and with anything else we try to deserialise an error message, and return generic error if we can't. The httpClient can be configured to handle redirects etc. but that's up to the calling code. If any reach this point there's nothing we can do (or is justified in doing right now anyway), thus treat as an error.