Ensure created classifiers are cleaned up
When running the smoke tests, the create classifier tests leave a
classifier called `test-classifier` hanging around, which then affects
the running of the delete classifier tests: if the create classifier
tests haven't run, the delete classifier test fails; if it has, then the
test succeeds.

Cleaning up the classifiers after the test requires the delete
classifier test to create the classifier before attempting to delete it.