Replace wildcard matching with regex
This takes a little bit of jiggery-pokery, hence the helper function.
Essentially, we have to escape the input string so that it is matched as
a literal string, and then append a `.*` to the end of the escaped
string to ensure that anything following the end of the literal string
is matched too.