Test that untrained classifiers are deleted
If a classifier is created but cannot be trained from the provided
samples archive file, delete the classifier so it appears to be an
atomic operation.

This is fairly involved, introducing a new ClassifierCommand interface
which is implemented by DeleteClassifier; this can then be mocked by
the new test. The NewDeleteClassifier() function needs to return a
ClassifiersCommand interface, which then has wider-reaching implications
for the rest of this area of the codebase.