Fix issue where classification results were mixed up
Go allocates memory just once for a variable like `job` in a for loop like `for job := range jobsChan {`. We were referencing this with `&job` and pushing that memory address onto the results channel. But before printing the results, this memory had been updated with a new job - resulting in incorrect filename  being output with each result.