Remove JWT parsing and refresh against API reponse
The API token endpoint returns an `expires_in` value which indicates how
long the token is valid for. By using this we can generate an "expires
at" value to use in the cache. The cache code is now much simpler than
it was.

This is a big change, as we have to change the public interface of
TokenRetriever. This has knock-on effects mandating the move of the auth
tests to their own package.