Fix duplicate error messages being printed
We had a mix of printing errors within the command and returning errors for printing in ch360.go. This led to some cases of duplicate error messages, including "context canceled" errors. Now the command never prints errors. It already only returns the first error, so the calling code can continue to print that error and everything's good.