Make DetailedErrorResponse.Error return detail if present
The rfc7807 problem responses can include detailed information about the
error. This change ensures it is used in preference to the 'title', in
the `Error` method of the DetailedErrorResponse type.

The `create extractor` command is also updated to use the `Error` method
in place of `Title`, directly, so attempting to create a duplicate
extractor will now result in an error message like this:

`Extractor creation failed with the following error: There is already an
extractor called 'test'. Please use another name for the new extractor.`